Se termino de implementar todas las funciones

+++ src/App.jsx
@@ -8,13 +8,16 @@ import axios from 'axios';
 function App() {
   const baseUrl = "http://localhost:3000/posts"
   const [table, setTable] = useState([]);
-
+  const [search, setSearch] = useState("");
   const deletePost = (event) => {
-    axios.delete(`http://localhost:3000/posts/${event.target.value}`);
+    if (window.confirm("Seguro que queres eleminar este post?")) {
+      axios.delete(`http://localhost:3000/posts/${event.target.value}`);
+    }
   }
 
   const addPost = (post) => {
     axios.post(baseUrl, post)
+    .then(() => alert("Se añadio el post"))
     .catch(error => {
       console.log(error);
     });
@@ -22,14 +25,17 @@ function App() {
 
   useEffect(() => {
     axios.get(baseUrl)
-      .then(json => setTable(json.data));
+      .then(json => setTable(json.data))
+      .catch(error => {
+        console.log(error);
+      });
   }, );
-
+  
   return (
     <>
-      <NavBar />
+      <NavBar setSearch={setSearch}/>
+      <Table data={table} deletePost={deletePost} search={search}/>
       <Form data={table} addPost={addPost}/>
-      <Table data={table} deletePost={deletePost} />
       <Footer />
     </>
   )
